+++ renderer/src/constants.js
@@ -3,8 +3,8 @@
  */
 
 export const APP_NAME = 'Quantro';
-export const APP_COMPANY = 'Quantro';
-export const APP_VERSION = '1.1.8';
+export const APP_COMPANY = 'Maze ERP';
+export const APP_VERSION = '1.1.9';
 export const PURCHASES_LABEL = 'Purchases';
 
 export const EMPTY_PRODUCT = {